- add typescript config - install react typescript dependencies - rename files
