+++ src/components/Header.tsx
@@ -1,8 +1,13 @@
-import { Link } from "gatsby"
-import PropTypes from "prop-types"
 import React from "react"
+import { Link } from "gatsby"
 
-const Header = ({ siteTitle }) => (
+const defaultProps = {
+  siteTitle: `View the source`,
+}
+
+type HeaderProps = { description: string } & typeof defaultProps
+
+const Header = ({ siteTitle }: HeaderProps) => (
   <header
     style={{
       background: `rebeccapurple`,
@@ -31,12 +36,5 @@ const Header = ({ siteTitle }) => (
   </header>
 )
 
-Header.propTypes = {
-  siteTitle: PropTypes.string,
-}
-
-Header.defaultProps = {
-  siteTitle: ``,
-}
-
 export default Header
+Header.defaultProps = defaultProps;
